Move DFP config to admin module
